simplify rollup config
